Move RequestPipeline to its own file.
