+++ src/adapters/TelnetAdapter.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import type { EvaluateContainer } from './TelnetAdapter';
-import { TelnetAdapter, RequestPipeline } from './TelnetAdapter';
+import { TelnetAdapter } from './TelnetAdapter';
 import * as dedent from 'dedent';
 import { HighLevelType } from './DebugProtocolAdapter';
 
@@ -195,66 +195,3 @@ describe('TelnetAdapter ', () => {
         });
     });
 });
-
-describe('RequestPipeline', () => {
-    let pipeline: RequestPipeline;
-    let socket = {
-        listeners: [],
-        messageQueue: [] as Array<string[]>,
-        addListener: function(eventName: string, listener: (data: Buffer) => void) {
-            this.listeners.push(listener);
-        },
-        //custom function for tests used to emit data to listeners
-        emit: function(data: string) {
-            const buffer = Buffer.from(data);
-            for (const listener of this.listeners) {
-                listener(buffer);
-            }
-        },
-        write: async function(text: string) {
-            //flush messages after getting data written
-            for (const messages of this.messageQueue) {
-                for (const message of messages) {
-                    await sleep(1);
-                    this.emit(message);
-                }
-            }
-        }
-    };
-
-    beforeEach(() => {
-        socket.listeners = [];
-        pipeline = new RequestPipeline(socket as any);
-        pipeline['isAtDebuggerPrompt'] = true;
-    });
-
-    it('handles split debugger prompt messages', async () => {
-        socket.messageQueue.push([
-            'some text Brightsc',
-            'ript Debugger>'
-        ]);
-        expect(
-            await pipeline.executeCommand('doSomething', true)
-        ).to.eql(
-            'some text Brightscript Debugger>'
-        );
-    });
-
-    it('handles debugger prompt separate from data', async () => {
-        socket.messageQueue.push([
-            'some text',
-            ' Brightscript Debugger>'
-        ]);
-        expect(
-            await pipeline.executeCommand('doSomething', true)
-        ).to.eql(
-            'some text Brightscript Debugger>'
-        );
-    });
-});
-
-function sleep(ms: number) {
-    return new Promise((resolve) => {
-        setTimeout(resolve, ms);
-    });
-}
