+++ src/util.ts
@@ -297,6 +297,12 @@ class Util {
     }
 
     private dnsCache = new Map<string, string>();
+
+    public sleep(ms: number) {
+        return new Promise((resolve) => {
+            setTimeout(resolve, ms);
+        });
+    }
 }
 
 export function defer<T>() {
